Improve fastpass performance.
Previously, fastpass could take ~10 minutes to process a particularly
large number of Pants targets with large classpaths. Now, it takes ~60
seconds to process the same set of targets. I would like it to run
faster, but it's not a high priority since there are lower hanging
fruits (like optimizing Pants export-dep-as-jar which uses 5 minutes for
the same target set.