Keep globs when translating to Bloop
Previously, Fastpass was simplifying globs such as `**/*.java` to source
roots in the Bloop configuration. This means that, if the source root
included Scala source files, they would be compiled, whereas they would
have been ignored by Pants.

This commit removes the logic that simplifies the globs, so that we keep
the globs when writing the Bloop configuration.