Look for `scala-library` in libraries
Previously, we would attempt to detect the scala version used by a
target by looking for the Scala compiler in the libraries. It turns out,
however, that the Scala compiler is not always used as a library in a
target, and Fastpass would use the Scala version defined in this build
(Fastpass' build!) as Scala version of the created target.

This commit changes this logic so that we look for the Scala library
instead.

It also bumps the Scala version so that we have something more recent by
default.