Allow specifying coursier path in env variables
Previously, the path to the coursier binary could be specified only on
the command line by passing `--coursier-binary` to `fastpass create`.
This solution, while working, offered a poor user experience on machines
that don't have access to the Internet, because users would need to
remember about this parameter.

This commit allows specifying the path to the coursier binary via
environment variable (`COURSIER_BINARY`). When this environment variable
is set, and it's value describes an existing file, then Fastpass will
not try to download coursier.