Trigger refresh when creating existing projects
Previously, calling `fastpass create` with a project that already
existed would cause an error and ask the user to call `fastpass refresh`
instead.

This behavior complicates the user experience by requiring the user to
keep track of their existing project, even though Fastpass can
accurately figure out that the project can simply be refreshed.

With this commit, Fastpass will trigger the refresh of an existing
project if it exists and was created with the same import mode and list
of targets.