Support targets that can't be read by buildozer
Some targets are created using macros, so that they cannot be even found
when parsing the buildfiles with buildozer. For these targets, which we
assume are not being modified like regular targets, we register only the
source files that Bazel knows of at the time of analysis, rather than
trying to guess a source glob.