Infer `--bazel` and translate target specs
Previously, users who wanted to import their targets using Bazel had to
specify `--bazel`. This commit improves Fastpass so that the flag can be
inferred from the targets that are passed to Fastpass.

If there's a target that's prefixed by `//`, or suffixed by `/...`, then
we know that the Bazel is enabled. In that case, we can help the user
translate their target specs:
 - passing a directory to Fastpass is interpreted as recursively import
   all targets in that directory.
 - `foo/bar:` is interpreted as `foo/bar:all` (ie. all targets in the
   directory, non-recursively.
 - `foo/bar::` is rewritten to `foo/bar/...`