Generate synthetic project root
Previously, Fastpass with Bazel would not generate synthetic project
roots that would group together projects that share the same root
directory, unlike Fastpass with Pants. This would lead to Fastpass
projects that are hard to navigate in IntelliJ.

This commit ports the logic from Fastpass for Pants to Fastpass for
Bazel.