Infer enclosing resources root
Previously, Fastpass would consider that resource targets are rooted at
their package. It turns out that this is not correct, and that we should
use the enclosing resources directory. Typically, this means the
outermost enclosing directory named "resources", or "java".

If set, strip the "resource_strip_prefix" prefix from the resources
path.