Version cache format
Previously, Fastpass would version the cache format with the Fastpass
version. This means that the cache would be invalidated whenever
there's a new release of Fastpass.

Since we don't expect the cache format to change often, and we have
entirely control over when it changes, it's better to explicitly version
the cache format.