Refine dependencies to build
Previously, Fastpass could request Bazel builds some platform-dependent
dependencies, which may fail if these dependencies don't exist on a
given platform.

Because our BUILD files are such that these dependencies can only appear
as transitive dependencies, it's safe not to explicitly specify that
these should be built, since they will be built along with their
dependent.

This commit changes the logic used to gather the targets to build as
part of the import, so that targets that are windows-only (determined by
looking at the tag `jvm_classifier=windows`) are excluded.

Because this change impacts the cached export, the cache key version is
bumped.