Set sources in project-root instead of resources
Previously, Fastpass would create synthetic "project roots", and mark
them as resource directories, so that they act as aggregator in
IntelliJ's file tree.

Unfortunately, this doesn't work anymore: Because Bloop adds the
resources to the classpath, IntelliJ sees that the directory is a class
directory rather than a source directory, and removes its content root.
As a result, the directory doesn't appear in IntelliJ's file view and
doesn't act as an aggregator anymore.

To fix this, we set the source globs in the synthetic project root to
`.DS_Store` and create the file at the root of the synthetic project.
This way, the sources have an existing file, which IntelliJ will hide
and Bloop won't compile.