Fix go to definition on Thrift symbols in IntelliJ
Previously, running `go to definition` on Thrift symbols in IntelliJ
would take the user to decompiled class files, because the source jars
was not associated with the binary library. The source jars was no
associated because it would be named `<lib name>.srcjar`, which IntelliJ
would not recognize as a source jar. To fix the issue, we rename
`<lib name>.srcjar` to `<lib name>-sources.jar`, which IntelliJ will
correctly associate with `<lib name>.jar`.