Avoid pulling in unnecessary java bootclasspath
Previously, Fastpass would build the classpath of Java targets using the
inputs of both the `Javac` and `Scalac` actions. In addition to the jars
appearing in the `Scalac` action, the `Javac` action has a jar named
`platformclasspath.jar` which will be put on the bootclasspath of Javac.
This additional jar would end up on the compilation classpath of the
targets, causing issues down the line (eg. wrong go to definition in
IntelliJ).

Because our Java targets are implemented using `scala_library`, we can
rely on the `Scalac` action alone: it contains all the necessary jars.
The inputs of `Javac` were added by mistake when working on a different
issue, and should never have been added in the first place.