Handle base directory conflicts between targets
Previously, Fastpass would not take special considerations when 2 or
more targets shared the same base directory. Unfortunately, IntelliJ
currently does not allow multiple targets to share the same base
directory. As a result, some targets may end up being hidden, source
files may be associated with the wrong target, etc.

With this patch, Fastpass will try to rebase targets with conflicting
base directories where possible. To do so, Fastpass looks at the sources
of the targets, and if they share a common root that is not yet the base
directory of another target, then Fastpass will use that common root as
the new target's base directory.

When the conflict cannot be automatically resolved, Fastpass shows a
warning message.