Support `java_sources` in `scala_library` targets
Previously, Fastpass would ignore the `java_sources` field that appears
in some `scala_library` targets. This field is used to denote that a circular
dependency between this target and the label(s) set in `java_sources`
exists. In order to compile a target that has `java_sources` set, the
sources of the labels that are set must be passed to the compiler along
with the sources of the target.