Collect inputs of Javac mnemonic
Previously, Fastpass would not collect the inputs of the `Javac`
mnemonic for Java targets, and would focus only on the `Scalac`
mnemonic. Unfortunately, it turns out that we cannot do this, because it
appears that `strict_deps` is not applied to Java compilation with
Bazel: their classpath will still contain all the transitive
dependencies.

In order to avoid breaking go to definition on JDK classes in the IDE,
we exclude the `platformclasspath` JAR from the classpath (Bloop will
provide the right one from the environment).