Fix globs rebasing
Previously, in some cases, Fastpass could not correctly adapt the source
globs after rebasing a target's root directory. This would happen, for
instance, in a target:

 - whose base directory is foo/bar, and
 - whose globs is "hello/**/*.scala", and
 - whose source files are ["foo/bar/hello/src/my/dir/A.scala"], and
 - which is rebased to "foo/bar/hello/src/my/dir".

Previously, Fastpass would try to remove the new base directory prefix
from the source globs, which wouldn't work in this case for instance.

With this patch, Fastpass will correctly rewrite the globs.